Upgrades Patch style
Notices that edavis10 uses a different style for patches which work even better
and look cleaner. This will move the search patch to the correct and better
style.
