+++ init.rb
@@ -17,8 +17,6 @@
 
 require 'redmine'
 
-require 'quick_jump_equipment_patch'
-
 Redmine::Plugin.register :redmine_equipment_status_viewer do
   name 'Redmine Equipment Status Viewer plugin'
   author 'Devin Weaver'
@@ -42,4 +40,10 @@ Redmine::Plugin.register :redmine_equipment_status_viewer do
     :if => Proc.new {
       User.current.allowed_to?(:view_equipment_assets, nil, :global => true)
     }
+end
+
+require 'dispatcher'
+Dispatcher.to_prepare :redmine_equipment_status_viewer do
+  require_dependency 'search_controller'
+  SearchController.send(:include, RedmineEquipmentStatusViewer::Patches::SearchControllerPatch)
 end
